Broke up the monolithic 'ApiResponse' and added convenience methods for calling the API
