+++ app/src/main/java/fr/free/nrw/commons/mwapi/response/EditApiResponse.java
@@ -1,0 +1,13 @@
+package fr.free.nrw.commons.mwapi.response;
+
+import com.google.gson.annotations.SerializedName;
+
+public class EditApiResponse {
+    @SerializedName("edit")
+    public EditResponse edit;
+
+    public class EditResponse {
+        @SerializedName("result")
+        public String result;
+    }
+}
